Split chaikin indicator into several files.
