+++ js/masters/indicators/chaikin.src.js
@@ -11,4 +11,4 @@
  * License: www.highcharts.com/license
  */
 'use strict';
-import '../../Stock/Indicators/ChaikinIndicator.js';
+import '../../Stock/Indicators/Chaikin/ChaikinIndicator.js';
