+++ ts/Stock/Indicators/Chaikin/ChaikinPoint.d.ts
@@ -1,0 +1,34 @@
+/* *
+ *
+ *  License: www.highcharts.com/license
+ *
+ *  !!!!!!! SOURCE GETS TRANSPILED BY TYPESCRIPT. EDIT TS FILE ONLY. !!!!!!!
+ *
+ * */
+
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type ChaikinIndicator from './ChaikinIndicator';
+import type EMAPoint from '../EMA/EMAPoint';
+
+/* *
+ *
+ *  Class
+ *
+ * */
+
+declare class ChaikinPoint extends EMAPoint {
+    public series: ChaikinIndicator;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default ChaikinPoint;
